python/requirements.txt
requests
protobuf
grpcio
grpcio-tools
graphqlclient